refactor item and shares to use shard widget
